+++ src/tracer/vec3.rs
@@ -2,9 +2,9 @@ use std::ops::{Add, Div, Mul, Neg, Sub};
 
 #[derive(Clone, Copy, Default, Debug, PartialEq)]
 pub struct Vec3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl Add for Vec3 {
@@ -31,10 +31,10 @@ impl Sub for Vec3 {
     }
 }
 
-impl Mul<f64> for Vec3 {
+impl Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, other: f64) -> Vec3 {
+    fn mul(self, other: f32) -> Vec3 {
         Vec3 {
             x: self.x * other,
             y: self.y * other,
@@ -43,10 +43,10 @@ impl Mul<f64> for Vec3 {
     }
 }
 
-impl Div<f64> for Vec3 {
+impl Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, other: f64) -> Vec3 {
+    fn div(self, other: f32) -> Vec3 {
         Vec3 {
             x: self.x / other,
             y: self.y / other,
@@ -68,7 +68,7 @@ impl Neg for Vec3 {
 }
 
 impl Vec3 {
-    pub fn new(x: f64, y: f64, z: f64) -> Vec3 {
+    pub fn new(x: f32, y: f32, z: f32) -> Vec3 {
         Vec3 { x: x, y: y, z: z }
     }
     pub fn zero() -> Vec3 {
@@ -79,7 +79,7 @@ impl Vec3 {
         }
     }
 
-    pub fn dot(A: Vec3, B: Vec3) -> f64 {
+    pub fn dot(A: Vec3, B: Vec3) -> f32 {
         A.x * B.x + A.y * B.y + A.z * B.z
     }
 
@@ -100,7 +100,7 @@ impl Vec3 {
         ])
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self.x * self.x + self.y * self.y + self.z * self.z).sqrt()
     }
